Setup C++ test environment, split io helpers from index_builder

+++ cpp/include/scholar_rank/utils/vbe.hpp
@@ -5,7 +5,7 @@
 
 constexpr int BUFFER_LIMIT = 8;
 
-size_t vbe_encode(int num, unsigned char buffer[]);
+size_t vbe_encode(unsigned long long num, unsigned char buffer[]);
 
 unsigned long long vbe_decode(unsigned char buffer[]);
